Refactoring: move part of util.h to geometry.h
