+++ src/test/helpers/CMakeLists.txt
@@ -1,6 +1,7 @@
 target_sources(BrickSimTests PRIVATE
         test_color.cpp
         test_fraction.cpp
+        test_geometry.cpp
         test_polygon_cliping.cpp
         test_util.cpp
         )
